+++ src/languages/registry.ts
@@ -3,6 +3,9 @@ import { cFamilyLanguages } from "./definitions/c-family";
 import { scriptFamilyLanguages } from "./definitions/script-family";
 import { styleLanguages } from "./definitions/styles";
 import { dotnetFamilyLanguages } from "./definitions/dotnet-family";
+import { curlyFamilyLanguages } from "./definitions/curly-family";
+import { hashFamilyLanguages } from "./definitions/hash-family";
+import { dashFamilyLanguages } from "./definitions/dash-family";
 
 const byId = new Map<string, LanguageDefinition>();
 
@@ -19,6 +22,9 @@ for (const def of [
   ...scriptFamilyLanguages,
   ...styleLanguages,
   ...dotnetFamilyLanguages,
+  ...curlyFamilyLanguages,
+  ...hashFamilyLanguages,
+  ...dashFamilyLanguages,
 ]) {
   register(def);
 }